vdk-structlog: put vdk init logs config behind flag
Why?

Configuring the logs in the vdk_initialize hook causes commands like
vdk server and vdk info to have the same verbose logging format as
data job logs.

What?

- Put the log configuration for vdk_initialize behind a config option
and disable it by default for local runs and enable it by default
for cloud runs. Configuring the logging format in vdk_initialize might
still be useful for cloud runs and we don't care about other vdk commands
output in cloud environments

- Move config classes and functions into separate file

How was this tested

Manually
CI/CD

What kind of change is this?
Feature/non-breaking
