+++ projects/vdk-plugins/vdk-structlog/README.md
@@ -23,11 +23,12 @@ pip install vdk-structlog
 
 | Name                          | Description                                                                                                                                                                                                                      | Example Value                                           | Possible Values                                                                                                                                                                                                                                                     |
 |-------------------------------|----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|---------------------------------------------------------|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-| use_structlog                 | Use the structlog logging config instead of using the one in vdk-core                                                                                                                                                            | "True"                                                  | "True", "False"                                                                                                                                                                                                                                                     |
+| use_structlog                 | Use the structlog logging config instead of using the one in vdk-core (True by default).                                                                                                                                         | "True"                                                  | "True", "False"                                                                                                                                                                                                                                                     |
 | structlog_metadata            | Configure the metadata that will be output along with the log message                                                                                                                                                            | "timestamp, level, logger_name, file_name, vdk_job_name | Any combination of the following: "timestamp, level, logger_name, file_name, line_number, function_name, vdk_job_name, vdk_step_name, vdk_step_type". Can be expanded by extra params and bound key-value pairs. See the bound logger examples for more information |
 | structlog_format              | Configure the logging output format. Available formats: json, console, ltsv                                                                                                                                                      | "console"                                               | "console", "json", "ltsv"                                                                                                                                                                                                                                           |
 | structlog_console_log_pattern | Custom format string for console logging, applied only when`logging_format` is 'console'. Overrides `logging_metadata`. Note: For config.ini, %-signs should be escaped by doubling, e.g. %(asctime)s should become %%(asctime)s | "%(asctime)s %(name)-12s %(levelname)-8s %(message)s"   | Any valid Python logging format string                                                                                                                                                                                                                              |
 | structlog_config_preset       | Choose a configuration preset. Any config options set together with the preset will override the preset options. Available presets: LOCAL, CLOUD.                                                                                | "CLOUD"                                                 | "console", "json", "ltsv"                                                                                                                                                                                                                                           |
+| structlog_format_init_logs    | Set to True to apply structlog formatting options to vdk initialization logs                                                                                                                                                     | "True"                                                  | "True", "False"                                                                                                                                                                                                                                                     |
 | log_level_module              | Configure the log level of different Python modules separately                                                                                                                                                                   | "a.b.c=INFO;foo.bar=ERROR"                              | Semicolon-separated list of pairs of Python module paths and log level labels                                                                                                                                                                                       |
 | syslog_host                   | Syslog host to which logs are emitted                                                                                                                                                                                            | "syslog.vmware.com"                                     | Any valid host name                                                                                                                                                                                                                                                 |
 | syslog_port                   | Syslog port used to emit logs                                                                                                                                                                                                    | 514                                                     | Any valid port number                                                                                                                                                                                                                                               |
